refactor: update structure
