--- youtube_client_app/src/app/components/shared/search/search-item/search-item.component.html
@@ -1,4 +1,0 @@
-<div>
-  <h3>Search-item</h3>
-  <p>{{searchItem.snippet.title}}</p>
-</div>
