+++ youtube_client_app/src/app/core/header/header.component.ts
