+++ youtube_client_app/src/app/core/header/profile/profile.component.ts
