+++ youtube_client_app/src/app/core/header/search-panel/search-panel.component.html
