+++ youtube_client_app/src/app/main-page/search/search-item/search-item.component.html
@@ -1,0 +1,4 @@
+<div>
+  <h3>Search-item</h3>
+  <p>{{title}}</p>
+</div>
